Move StringTokenizer to base/tools and resolve includes
