+++ base/src/sgpp/base/tools/StringTokenizer.hpp
@@ -11,11 +11,12 @@
 #include <vector>
 
 namespace sgpp {
-namespace datadriven {
+namespace base {
 
 class StringTokenizer {
  public:
-  static void tokenize(const std::string &s, const char *delim, std::vector<std::string> &v) {
+  static void tokenize(const std::string &s, const char *delim,
+                       std::vector<std::string> &v) {
     size_t pos = 0;
     size_t start = 0;
 
@@ -29,5 +30,5 @@ class StringTokenizer {
     }
   }
 };
-} /* namespace datadriven */
+} /* namespace base */
 } /* namespace sgpp */
