+++ datadriven/src/sgpp/datadriven/algorithm/DBMatOfflineFactory.cpp
@@ -11,10 +11,11 @@
 #include <sgpp/datadriven/algorithm/DBMatOfflineEigen.hpp>
 #include <sgpp/datadriven/algorithm/DBMatOfflineLU.hpp>
 #include <sgpp/datadriven/algorithm/DBMatOfflineOrthoAdapt.hpp>
-#include <sgpp/datadriven/datamining/base/StringTokenizer.hpp>
+#include <sgpp/base/tools/StringTokenizer.hpp>
 
 #include <string>
 #include <vector>
+
 
 namespace sgpp {
 namespace datadriven {
@@ -81,7 +82,7 @@ DBMatOffline* DBMatOfflineFactory::buildFromFile(const std::string& fileName) {
 
   std::cout << str;
   std::vector<std::string> tokens;
-  StringTokenizer::tokenize(str, ",", tokens);
+  sgpp::base::StringTokenizer::tokenize(str, ",", tokens);
   std::cout << "tokens: ";
   for (auto& item : tokens) {
     std::cout << item << ",";
